scripts/requirements.txt
httpx
beautifulsoup4
pdfplumber
supabase
tqdm
python-dotenv
google-genai
pydantic
